requirements.txt
# ---------------------------------------------------------
# Core ML and vision dependencies (GPU-enabled)
# ---------------------------------------------------------
torch==2.7.1                   # GPU-enabled PyTorch (CUDA 11.8)
torchvision==0.22.1            # GPU-enabled torchvision
torchaudio==2.7.1              # Optional but safe to include
ultralytics>=8.1.0             # YOLOv11 support
opencv-python>=4.5.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.64.0
PyYAML>=6.0

# ---------------------------------------------------------
# Roboflow integration and data handling
# ---------------------------------------------------------
roboflow>=1.0.0
requests>=2.28.0
gdown

# ---------------------------------------------------------
# Web deployment (Flask + CORS)
# ---------------------------------------------------------
flask>=2.2.0
flask-cors>=3.1.0

# ---------------------------------------------------------
# Development and logging utilities
# ---------------------------------------------------------
psutil
gitpython>=3.1.30
setuptools>=65.5.1
dill==0.3.5.1
from-root==1.0.2

# ---------------------------------------------------------
# Notebook and visualization (optional)
# ---------------------------------------------------------
notebook>=6.5.0
pandas>=1.3.0
seaborn>=0.11.0

# ---------------------------------------------------------
# Local package reference
# ---------------------------------------------------------
-e .